+++ src/parser/inline/code.test.ts
@@ -14,6 +14,7 @@ describe('Unit: parser/inline/code', () => {
       assert.deepStrictEqual(inspect(parser('` `` `')), void 0);
       assert.deepStrictEqual(inspect(parser('`\n`')), void 0);
       assert.deepStrictEqual(inspect(parser('`a\nb`')), void 0);
+      assert.deepStrictEqual(inspect(parser('`a\\\nb`')), void 0);
       assert.deepStrictEqual(inspect(parser('a`b`')), void 0);
     });