+++ src/parser/block/pretext.test.ts
@@ -30,7 +30,10 @@ describe('Unit: parser/block/pretext', () => {
 
     it('attribute', () => {
       assert.deepStrictEqual(inspect(parser('```abc\na\n```')), [['<pre class="language-abc" data-lang="abc">a</pre>'], '']);
+      assert.deepStrictEqual(inspect(parser('```abc \na\n```')), [['<pre class="language-abc" data-lang="abc">a</pre>'], '']);
+      assert.deepStrictEqual(inspect(parser('```a b.c\n\n```')), [['<pre class="language-a" data-lang="a" data-file="b.c"></pre>'], '']);
       assert.deepStrictEqual(inspect(parser('```a  b.c  \n\n```')), [['<pre class="language-a" data-lang="a" data-file="b.c"></pre>'], '']);
+      assert.deepStrictEqual(inspect(parser('```a b c\n\n```')), [['<pre class="language-a" data-lang="a" data-file="b"></pre>'], '']);
     });
 
   });