+++ src/parser/block/pretext.ts
@@ -5,16 +5,17 @@ import { UnescapableSourceParser } from '../source';
 
 type SubParsers = [UnescapableSourceParser];
 
-const syntax = /^(`{3,})([a-z]*)(?:[^\S\n]+([0-9a-zA-Z_\-.]+))?[^\S\n]*\n(?:[^\n]*\n)+?\1[^\S\n]*(?=\n|$)/;
+const syntax = /^(`{3,})[^\n]*\n(?:[^\n]*\n)+?\1[^\S\n]*(?=\n|$)/;
 
 export const pretext: PreTextParser = verifyBlockEnd(function (source: string): Result<HTMLPreElement, SubParsers> {
   if (!source.startsWith('```')) return;
-  const [whole, , lang, filename = ''] = source.match(syntax) || ['', '', ''];
+  const [whole] = source.match(syntax) || [''];
   if (!whole) return;
+  const [, lang, filename = ''] = source.split('\n', 1)[0].match(/^(?:`{3,})([a-z]*)(?:\s+([0-9a-zA-Z_\-.]+))?(?=\s|$)/);
   const el = document.createElement('pre');
   if (lang) {
     void el.setAttribute('class', `language-${lang.toLowerCase()}`);
-    void el.setAttribute('data-lang', `${lang.toLowerCase()}`);
+    void el.setAttribute('data-lang', lang.toLowerCase());
   }
   if (filename) {
     void el.setAttribute('data-file', filename);